+++ src/Symfony/Component/Cache/Adapter/DoctrineAdapter.php
@@ -18,6 +18,11 @@ class DoctrineAdapter extends AbstractAdapter
 {
     use DoctrineTrait;
 
+    /**
+     * @param CacheProvider $provider
+     * @param string        $namespace
+     * @param int           $defaultLifetime
+     */
     public function __construct(CacheProvider $provider, $namespace = '', $defaultLifetime = 0)
     {
         parent::__construct('', $defaultLifetime);