+++ src/Symfony/Component/Cache/Simple/MemcachedCache.php
@@ -19,6 +19,11 @@ class MemcachedCache extends AbstractCache
 
     protected $maxIdLength = 250;
 
+    /**
+     * @param \Memcached $client
+     * @param string     $namespace
+     * @param int        $defaultLifetime
+     */
     public function __construct(\Memcached $client, $namespace = '', $defaultLifetime = 0)
     {
         $this->init($client, $namespace, $defaultLifetime);