+++ src/Symfony/Component/DependencyInjection/Dumper/PhpDumper.php
@@ -121,7 +121,6 @@ class PhpDumper extends Dumper
             'debug' => true,
         ), $options);
 
-        $this->classResources = array();
         $this->initializeMethodNamesMap($options['base_class']);
 
         $this->docStar = $options['debug'] ? '*' : '';